+++ packages/framework/src/Concerns/HasPageMetadata.php
@@ -17,7 +17,7 @@ trait HasPageMetadata
 {
     public function getCanonicalUrl(): string
     {
-        return Hyde::qualifiedUrl(Hyde::pageLink($this->getCurrentPagePath().'.html'));
+        return Hyde::url(Hyde::pageLink($this->getCurrentPagePath().'.html'));
     }
 
     public function getDynamicMetadata(): array
@@ -29,14 +29,14 @@ trait HasPageMetadata
         }
 
         if ($this->canUseSitemapLink()) {
-            $array[] = '<link rel="sitemap" type="application/xml" title="Sitemap" href="'.Hyde::qualifiedUrl('sitemap.xml').'" />';
+            $array[] = '<link rel="sitemap" type="application/xml" title="Sitemap" href="'.Hyde::url('sitemap.xml').'" />';
         }
 
         if ($this->canUseRssFeedlink()) {
             $array[] = '<link rel="alternate" type="application/rss+xml" title="'
             .RssFeedService::getTitle()
             .' RSS Feed" href="'
-            .Hyde::qualifiedUrl(RssFeedService::getDefaultOutputFilename())
+            .Hyde::url(RssFeedService::getDefaultOutputFilename())
             .'" />';
         }