+++ packages/framework/src/HydeKernel.php
@@ -198,7 +198,7 @@ class HydeKernel implements HydeKernelContract
     /**
      * Return a qualified URI path, if SITE_URL is set in .env, else return false.
      *
-     * @deprecated v0.53.0-beta - Use Hyde::qualifiedUrl() or Hyde::hasSiteUrl() instead.
+     * @deprecated v0.53.0-beta - Use Hyde::url() or Hyde::hasSiteUrl() instead.
      *
      * @param  string  $path  optional relative path suffix. Omit to return base url.
      * @return string|false
@@ -229,7 +229,7 @@ class HydeKernel implements HydeKernelContract
      *
      * @throws \Exception If no site URL is set and no default is provided
      */
-    public function qualifiedUrl(string $path = '', ?string $default = null): string
+    public function url(string $path = '', ?string $default = null): string
     {
         if ($this->hasSiteUrl()) {
             return rtrim(rtrim(config('site.url'), '/').'/'.(trim($path, '/') ?? ''), '/');